# commerceue-app/WebContent/swagger_v3.0.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"API Extensions for IBM Digital Commerce customizations.",
      "version":"3.0",
      "title":"IBM Digital Commerce API Extensions"
   },
   "basePath":"/commerceue/extension",
...
   "paths":{  
...
      "/person/person_register_post":{  
         "post":{  
            "tags":[  
               "person"
            ],
            "summary":"Customize actions to perform after registering a user.",
            "description":"| Command: com.ibm.commerce.usermanagement.commands.UserRegistrationAddCmd |\n|----------|\n| UserRegistrationAddCmd is a controller command which is used to register a new shopper|",
            "operationId":"personRegisterPost",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"personRegisterPost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/PersonPostUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/PersonPostUEResponse"
                  }
               },
               "400":{  
                  "description":"Application error"
               },
               "500":{  
                  "description":"System error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.usermanagement.commands.UserRegistrationAddCmd"
            }
         }
      },
      "/person/person_register_pre":{  
         "post":{  
            "tags":[  
               "person"
            ],
            "summary":"Customize actions to perform before registering a user.",
            "description":"| Command: com.ibm.commerce.usermanagement.commands.UserRegistrationAddCmd |\n|----------|\n| UserRegistrationAddCmd is a controller command which is used to register a new shopper|",
            "operationId":"personRegisterPre",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"personRegisterPre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/PersonPreUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/PersonPreUEResponse"
                  }
               },
               "400":{  
                  "description":"Application error"
               },
               "500":{  
                  "description":"System error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.usermanagement.commands.UserRegistrationAddCmd"
            }
         }
      },
      "/person/person_update_post":{  
         "post":{  
            "tags":[  
               "person"
            ],
            "summary":"Customize actions to perform after updating a user.",
            "description":"| Command: com.ibm.commerce.usermanagement.commands.UserRegistrationUpdateCmd |\n|----------|\n| UserRegistrationUpdateCmd is a controller command which is used to update existing shopper|",
            "operationId":"personUpdatePost",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"personUpdatePost UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/PersonPostUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/PersonPostUEResponse"
                  }
               },
               "400":{  
                  "description":"Application error"
               },
               "500":{  
                  "description":"System error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.usermanagement.commands.UserRegistrationUpdateCmd"
            }
         }
      },
      "/person/person_update_pre":{  
         "post":{  
            "tags":[  
               "person"
            ],
            "summary":"Customize actions to perform before updating a user.",
            "description":"| Command: com.ibm.commerce.usermanagement.commands.UserRegistrationUpdateCmd |\n|----------|\n| UserRegistrationUpdateCmd is a controller command which is used to update existing shopper|",
            "operationId":"personUpdatePre",
            "consumes":[  
               "application/json"
            ],
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"personUpdatePre UE input",
                  "description":"The UE Request String",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/PersonPreUERequest"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/PersonPreUEResponse"
                  }
               },
               "400":{  
                  "description":"Application error"
               },
               "500":{  
                  "description":"System error"
               }
            },
            "x-data-command":{  
               "Command":"com.ibm.commerce.usermanagement.commands.UserRegistrationUpdateCmd"
            }
         }
      },
...
   },
   "definitions":{  
...
      "ContactInfoExternalIdentifier":{  
         "type":"object",
         "properties":{  
            "contactInfoNickName":{  
               "type":"string"
            },
            "personIdentifier":{  
               "$ref":"#/definitions/PersonIdentifier"
            },
            "organizationIdentifier":{  
               "$ref":"#/definitions/OrganizationIdentifier"
            }
         }
      },
...
      "Address":{  
         "type":"object",
         "properties":{  
            "addressLine":{  
               "type":"array",
               "items":{  
                  "type":"string"
               }
            },
            "internalOfficeAddress":{  
               "type":"string"
            },
            "city":{  
               "type":"string"
            },
            "stateOrProvinceName":{  
               "type":"string"
            },
            "country":{  
               "type":"string"
            },
            "postalCode":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "type":{  
               "type":"string"
            },
            "primary":{  
               "type":"boolean",
               "default":false
            }
         }
      },
...
      "CheckoutProfile":{  
         "type":"object",
         "properties":{  
            "paymentInfo":{  
               "$ref":"#/definitions/PaymentInfo"
            },
            "shippingInfo":{  
               "$ref":"#/definitions/ShippingInfo"
            },
            "lastUpdate":{  
               "type":"string",
               "format":"date-time"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "ContactInfo":{  
         "type":"object",
         "properties":{  
            "contactInfoIdentifier":{  
               "$ref":"#/definitions/ContactInfoIdentifier"
            },
            "contactName":{  
               "$ref":"#/definitions/PersonName"
            },
            "address":{  
               "$ref":"#/definitions/Address"
            },
            "telephone1":{  
               "$ref":"#/definitions/Telephone"
            },
            "telephone2":{  
               "$ref":"#/definitions/Telephone"
            },
            "bestCallingTime":{  
               "type":"string"
            },
            "emailAddress1":{  
               "$ref":"#/definitions/EmailAddress"
            },
            "emailAddress2":{  
               "$ref":"#/definitions/EmailAddress"
            },
            "fax1":{  
               "$ref":"#/definitions/Fax"
            },
            "fax2":{  
               "$ref":"#/definitions/Fax"
            },
            "mobilePhone1":{  
               "$ref":"#/definitions/MobilePhone"
            },
            "organizationName":{  
               "type":"string"
            },
            "organizationUnitName":{  
               "type":"string"
            },
            "geographicalShippingCode":{  
               "type":"string"
            },
            "geographicalTaxCode":{  
               "type":"string"
            },
            "attributes":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "language":{  
               "type":"string"
            }
         }
      },
...
      "PaymentInfo":{  
         "type":"object",
         "properties":{  
            "billingAddress":{  
               "$ref":"#/definitions/ContactInfo"
            },
            "paymentMethod":{  
               "$ref":"#/definitions/PaymentMethod"
            },
            "protocolData":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ProtocolData"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "EmailAddress":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "BusinessProfile":{  
         "type":"object",
         "properties":{  
            "employeeID":{  
               "type":"string"
            },
            "employeeType":{  
               "type":"string"
            },
            "departmentNumber":{  
               "type":"string"
            },
            "manager":{  
               "type":"string"
            },
            "secretary":{  
               "type":"string"
            },
            "attributes":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "Person":{  
         "type":"object",
         "properties":{  
            "personIdentifier":{  
               "$ref":"#/definitions/PersonIdentifier"
            },
            "parentOrganizationIdentifier":{  
               "$ref":"#/definitions/OrganizationIdentifier"
            },
            "credential":{  
               "$ref":"#/definitions/Credential"
            },
            "personalProfile":{  
               "$ref":"#/definitions/PersonalProfile"
            },
            "businessProfile":{  
               "$ref":"#/definitions/BusinessProfile"
            },
            "contactInfo":{  
               "$ref":"#/definitions/ContactInfo"
            },
            "addressBook":{  
               "$ref":"#/definitions/AddressBook"
            },
            "contextAttribute":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/Attribute"
               }
            },
            "attributes":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "checkoutProfile":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/CheckoutProfile"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "registrationStatus":{  
               "type":"string"
            },
            "registrationApprovalStatus":{  
               "type":"string"
            },
            "passwordExpired":{  
               "type":"boolean",
               "default":false
            },
            "accountStatus":{  
               "type":"string"
            },
            "profile":{  
               "type":"string",
               "enum":[  
                  "CONSUMER",
                  "BUSINESS"
               ]
            }
         }
      },
...
      "Description":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "language":{  
               "type":"string"
            }
         }
      },
...
      "Attribute":{  
         "type":"object",
         "properties":{  
            "name":{  
               "type":"string"
            },
            "attributeValue":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/AttributeValue"
               }
            }
         }
      },
...
      "PersonPreUEResponse":{  
         "type":"object",
         "properties":{  
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "person":{  
               "description":"The person pojo object",
               "$ref":"#/definitions/Person"
            }
         },
         "description":"This is the pre UE response for the user registration controller commands"
      },
...
      "IncomeAmount":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"integer",
               "format":"int32"
            },
            "currency":{  
               "type":"string"
            }
         }
      },
...
      "ReceiveEmailPreference":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"boolean",
               "default":false
            },
            "storeID":{  
               "type":"string"
            }
         }
      },
...
      "Credential":{  
         "type":"object",
         "properties":{  
            "logonID":{  
               "type":"string"
            },
            "password":{  
               "type":"string"
            },
            "securityHint":{  
               "$ref":"#/definitions/SecurityHint"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "PersonPreUERequest":{  
         "type":"object",
         "properties":{  
            "contextData":{  
               "$ref":"#/definitions/ContextData"
            },
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "person":{  
               "$ref":"#/definitions/Person"
            }
         }
      },
...
      "Locale":{  
         "type":"object",
         "properties":{  
            "language":{  
               "type":"string"
            },
            "script":{  
               "type":"string"
            },
            "country":{  
               "type":"string"
            },
            "variant":{  
               "type":"string"
            },
            "extensionKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleAttributes":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "unicodeLocaleKeys":{  
               "type":"array",
               "uniqueItems":true,
               "items":{  
                  "type":"string"
               }
            },
            "iso3Language":{  
               "type":"string"
            },
            "iso3Country":{  
               "type":"string"
            },
            "displayLanguage":{  
               "type":"string"
            },
            "displayScript":{  
               "type":"string"
            },
            "displayCountry":{  
               "type":"string"
            },
            "displayVariant":{  
               "type":"string"
            },
            "displayName":{  
               "type":"string"
            }
         }
      },
...
      "AttributeValue":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"array",
               "items":{  
                  "type":"string"
               }
            },
            "storeID":{  
               "type":"string"
            }
         }
      },
...
      "PersonalProfile":{  
         "type":"object",
         "properties":{  
            "preferredCurrency":{  
               "type":"string"
            },
            "preferredLanguage":{  
               "type":"string"
            },
            "displayName":{  
               "type":"string"
            },
            "description":{  
               "type":"string"
            },
            "photoURI":{  
               "type":"string"
            },
            "preferredCommunication":{  
               "type":"string"
            },
            "preferredDelivery":{  
               "type":"string"
            },
            "receiveSMSNotification":{  
               "type":"boolean",
               "default":false
            },
            "receiveEmailPreference":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ReceiveEmailPreference"
               }
            },
            "receiveSMSPreference":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ReceiveSMSPreference"
               }
            },
            "gender":{  
               "type":"string"
            },
            "income":{  
               "$ref":"#/definitions/IncomeAmount"
            },
            "maritalStatus":{  
               "type":"string"
            },
            "numberOfChildren":{  
               "type":"integer",
               "format":"int32"
            },
            "householdSize":{  
               "type":"integer",
               "format":"int32"
            },
            "companyName":{  
               "type":"string"
            },
            "hobbies":{  
               "type":"string"
            },
            "dateOfBirth":{  
               "type":"string",
               "format":"date-time"
            },
            "registrationDateTime":{  
               "type":"string",
               "format":"date-time"
            },
            "lastUpdate":{  
               "type":"string",
               "format":"date-time"
            },
            "attributes":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
      "MobilePhone":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "country":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "PersonIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "distinguishedName":{  
               "type":"string"
            },
            "externalIdentifier":{  
               "$ref":"#/definitions/PersonExternalIdentifier"
            }
         }
      },
      "AddressBook":{  
         "type":"object",
         "properties":{  
            "contact":{  
               "type":"array",
               "items":{  
                  "$ref":"#/definitions/ContactInfo"
               }
            }
         }
      },
...
      "ShippingMode":{  
         "type":"object",
         "properties":{  
            "shippingModeIdentifier":{  
               "$ref":"#/definitions/ShippingModeIdentifier"
            },
            "description":{  
               "$ref":"#/definitions/Description"
            },
            "trackingURL":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
      "ContextData":{  
         "type":"object",
         "properties":{  
            "baseContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "accessControlContext":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            },
            "locale":{  
               "$ref":"#/definitions/Locale"
            }
         }
      },
      "ContactInfoIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "externalIdentifier":{  
               "$ref":"#/definitions/ContactInfoExternalIdentifier"
            }
         }
      },
      "ShippingInfo":{  
         "type":"object",
         "properties":{  
            "shippingAddress":{  
               "$ref":"#/definitions/ContactInfo"
            },
            "shippingMode":{  
               "$ref":"#/definitions/ShippingMode"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "PersonPostUERequest":{  
         "type":"object",
         "properties":{  
            "contextData":{  
               "$ref":"#/definitions/ContextData"
            },
            "commandInputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "commandOutputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "person":{  
               "$ref":"#/definitions/Person"
            }
         }
      },
...
      "PersonExternalIdentifier":{  
         "type":"object",
         "properties":{  
            "identifier":{  
               "type":"string"
            }
         }
      },
      "OrganizationIdentifier":{  
         "type":"object",
         "properties":{  
            "uniqueID":{  
               "type":"string"
            },
            "distinguishedName":{  
               "type":"string"
            }
         }
      },
...
      "PaymentMethod":{  
         "type":"object",
         "properties":{  
            "paymentMethodName":{  
               "type":"string"
            },
            "description":{  
               "$ref":"#/definitions/Description"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "Telephone":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            },
            "type":{  
               "type":"string"
            },
            "publish":{  
               "type":"boolean",
               "default":false
            }
         }
      },
      "UserData":{  
         "type":"object",
         "properties":{  
            "userDataField":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"string"
               }
            }
         }
      },
...
      "ReceiveSMSPreference":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"boolean",
               "default":false
            },
            "storeID":{  
               "type":"string"
            }
         }
      },
...
      "ProtocolData":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "name":{  
               "type":"string"
            }
         }
      },
...
      "PersonName":{  
         "type":"object",
         "properties":{  
            "personTitle":{  
               "type":"string"
            },
            "businessTitle":{  
               "type":"string"
            },
            "lastName":{  
               "type":"string"
            },
            "firstName":{  
               "type":"string"
            },
            "middleName":{  
               "type":"string"
            }
         }
      },
...
      "PersonPostUEResponse":{  
         "type":"object",
         "properties":{  
            "commandOutputs":{  
               "type":"object",
               "additionalProperties":{  
                  "type":"object"
               }
            },
            "person":{  
               "description":"The person pojo object",
               "$ref":"#/definitions/Person"
            }
         },
         "description":"This is the post UE response for the user registration controller commands"
      },
...
      "SecurityHint":{  
         "type":"object",
         "properties":{  
            "question":{  
               "type":"string"
            },
            "answer":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
      "Fax":{  
         "type":"object",
         "properties":{  
            "value":{  
               "type":"string"
            },
            "userData":{  
               "$ref":"#/definitions/UserData"
            }
         }
      },
...
   }
}